fix(cfs): might schedule duplicated tasks of a single task when multicore is on

+++ NoAxiom/kernel/src/syscall/syscall.rs
@@ -56,7 +56,7 @@ impl<'a> Syscall<'a> {
             SYS_EXIT => self.sys_exit(args[0]),
             SYS_CLONE => self.sys_fork(args[0], args[1], args[2], args[3], args[4]),
             SYS_EXECVE => self.sys_exec(args[0], args[1], args[2]).await,
-            SYS_WAIT4 => self.sys_wait4(args[0], args[1], args[2], args[3]).await,
+            SYS_WAIT4 => self.sys_wait4(args[0] as isize, args[1], args[2], args[3]).await,
             SYS_GETPID => self.sys_getpid(),
             SYS_GETPPID => self.sys_getppid(),
